% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GigaPRS_Step2.R
\name{GigaPRS_Step2}
\alias{GigaPRS_Step2}
\title{GigaPRS Step 2}
\usage{
GigaPRS_Step2(
  bedstem = NULL,
  scores = NULL,
  multiscores = NULL,
  outstem = NULL,
  allowMissing = FALSE,
  extractfile = NULL,
  keepfile = NULL,
  phenofile = NULL,
  saveCounts = FALSE
)
}
\arguments{
\item{bedstem}{The stem of the genotype files (e.g., if the genotypes are stored in file.bed, file.bim and file.fam, use bedstem="file")}

\item{scores}{A file containing single-ancestry SNP effect sizes (e.g., the output from running GigaPRS Step 1 with one focal ancestry)}

\item{multiscores}{A file containing multi-ancestry SNP effect sizes  (e.g., the output from running GigaPRS Step 1 with multiple focal ancestries)}

\item{outstem}{The desired prefix for the output files}

\item{allowMissing}{(TRUE or FALSE; default FALSE) Whether to allow missing genotypes for some score predictors}

\item{extractfile}{(Optional) A file containing a subset of predictors}

\item{keepfile}{(Optional) A file containing a subset of samples}

\item{phenofile}{(Optional) A file containing phenotypes for one or more samples}

\item{saveCounts}{(TRUE or FALSE; default FALSE) Whether to save number of non-missing values for each sample}
}
\description{
This function uses genotypes and SNP effect sizes to compute PRS for target individuals, possibly taking into account inferred ancestry
}
\details{
You need to provide genotypes (in PLINK 1 binary format) and either a single-ancestry or multi-ancestry scorefile
}
\examples{
#These examples use the gwas results that come with the GigaPRS package
ex_gwas_file=system.file("extdata", "ex_gwas.txt.gz", package="MegaPRS")

#Example 1 - Provide only the two required arguments
format_sumstats(gwasfile=ex_gwas_file, outstem="ex_out")

#Example 2 - Run the function based on the on-screen instructions from Example 1
format_sumstats(gwasfile=ex_gwas_file, outstem="ex_out", NameCol="Predictor", A1Col="A1", A2Col="A2", EffectCol="Beta", SECol="SE", nCol="n", FreqCol="A1Freq")
}
